- Fix bug in checking for number operands. - Change return type of clock() builtin to float as Lox only supports float in its "number" type .

+++ lox/callable.py
@@ -19,7 +19,7 @@ class Callable(ABC):
 
 class Clock(Callable):
     def call(self, interpreter, args: List[Any]):
-        return int(time.time())
+        return float(time.time())
 
     def arity(self) -> int:
         return 0
